third_party/blink/web_tests/platform/generic/paint/invalidation/svg/nested-embedded-svg-size-changes-expected.txt

This test checks that intrinsic size changes of an embedded SVG cause immediate updates of the object that embeds it.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Check initial <object> size, before external resource loaded
PASS document.defaultView.getComputedStyle(object1).width is "400px"
PASS document.defaultView.getComputedStyle(object1).height is "200px"


Check initial <object> size, after external resource loaded - should not change.
PASS document.defaultView.getComputedStyle(object1).width is "400px"
PASS document.defaultView.getComputedStyle(object1).height is "200px"


Check final <object> size, after resizing finished
PASS document.defaultView.getComputedStyle(object1).width is "400px"
PASS document.defaultView.getComputedStyle(object1).height is "200px"


Test passed if you see two green rectangles
PASS successfullyParsed is true

TEST COMPLETE
